Allow for typehinting of method signatures in deftype
Previously, protocols used typehints when generating interfaces, but
deftype discarded typehints when generating the code that implemented
the intefaces.  This would cause the compiler to barf, saying that it
couldn't find the matching method for a variadic method with
typehints.

The test fails to compile without the change to core_deftype.clj and
passes with the change.