Make APersistentVector.iterator slightly more efficient
Since nth on a PersistentVector has running time O(log_32(n)), and
calling next on a chunked sequence costs O(log_32(n)) 1/32 of the
time, it should be a bit more efficient to use sequences than
repeatedly calling nth.

Since calling next on a sequence causes a new object to be created,
there may be some values of n for which log_32(n) is less than the
number of operations needed for object creation.